requirements.txt
django
djangorestframework
djangorestframework-gis
django-cors-headers
psycopg2